pyproject.toml
[project]
name = "crawler"
version = "1.0.0"
description = "Web Crawler with NLP Processing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.13.5",
    "celery>=5.5.3",
    "fastapi>=0.116.1",
    "flower>=2.0.1",
    "jieba>=0.42.1",
    "loguru>=0.7.3",
    "lxml>=6.0.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "playwright>=1.55.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
    "scrapy>=2.13.3",
    "sqlalchemy>=2.0.43",
    "transformers>=4.56.0",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
    "ipython>=8.18.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
